User deletion changed, instead of being deleted a flag is set
Almost no request works with ids of deleted users, the 2 exceptions are
/login/userId/:username and /login/username/:userId.
Rooms now have a deletedUsers field to track users that deleted their
accounts while in this room, maybe a bit unnecessary.
Note: Rooms will be deleted when there are no more NON-DELETED users in
it.
Fix: when a room is deleted, posts are correctly deleted.